+++ src/component/sub-modules/base/src/box.js
@@ -156,6 +156,21 @@ KISSY.add('component/base/box', function () {
         },
 
         /**
+         * content element of current component.
+         * usually it is same with el.
+         * except eg: overlay, submenu
+         * @type {KISSY.NodeList}
+         * @readonly
+         * @property contentEl
+         */
+        /**
+         * @ignore
+         */
+        contentEl: {
+            view:1
+        },
+
+        /**
          * archor element where component append to
          * @cfg {KISSY.NodeList} render
          */
@@ -164,18 +179,6 @@ KISSY.add('component/base/box', function () {
          */
         render: {
             view: 1
-        },
-
-        /**
-         * component's visibleMode,use css "display" or "visibility" to show this component
-         * @cfg {String} visibleMode
-         */
-        /**
-         * @ignore
-         */
-        visibleMode: {
-            view: 1,
-            value: 'display'
         },
 
         /**