+++ src/menubutton/src/base.js
@@ -15,11 +15,10 @@ KISSY.add("menubutton/base", function (S, Node, Button, MenuButtonRender, Menu, 
             }
         };
     /**
-     * @class
      * A menu button component, consist of a button and a drop down popup menu.
      * xclass: 'menu-button'.
-     * @name MenuButton
-     * @extends Button
+     * @class KISSY.MenuButton
+     * @extends KISSY.Button
      */
     var MenuButton = Button.extend([Component.DecorateChild],
         /**
@@ -130,7 +129,7 @@ KISSY.add("menubutton/base", function (S, Node, Button, MenuButtonRender, Menu, 
             /**
              * Remove a existing menu item from drop down menu.
              * @param c {KISSY.Menu.Item} Existing menu item.
-             * @param [destroy] {Boolean} Whether destroy removed menu item.
+             * @param [destroy=true] {Boolean} Whether destroy removed menu item.
              */
             removeItem: function (c, destroy) {
                 /**
@@ -182,7 +181,7 @@ KISSY.add("menubutton/base", function (S, Node, Button, MenuButtonRender, Menu, 
             decorateChildrenInternal: function (UI, el) {
                 // 不能用 display:none , menu 的隐藏是靠 visibility
                 // eg: menu.show(); menu.hide();
-                el.css("visibility", "hidden").prependTo(el[0].ownerDocument.body);
+                el.css("top", "-9999").prependTo(el[0].ownerDocument.body);
                 var self = this;
                 self.setInternal("menu",
                     Component.DecorateChild.prototype.decorateChildrenInternal.call(self, UI, el, self.get('menu')));