+++ src/scrollview/sub-modules/plugin/scrollbar/src/control.js
@@ -195,7 +195,7 @@ KISSY.add('scrollview/plugin/scrollbar/control', function (S, Event, DD, Compone
                 self._startHideTimer();
             }
         },
-
+        // percentage matters!
         afterScrollChange: function (e) {
             // only show when scroll
             var self = this,
@@ -320,9 +320,6 @@ KISSY.add('scrollview/plugin/scrollbar/control', function (S, Event, DD, Compone
             },
             focusable: {
                 value: false
-            },
-            visibleMode: {
-                value: 'visibility'
             },
             xrender: {
                 value: ScrollBarRender